pjt_07/PJT-07/movies/serializers/reviews.py
from rest_framework.serializers import ModelSerializer
from ..models import Review, Movie


class ReviewListSerializer(ModelSerializer):
    class MovieForReview(ModelSerializer):
        class Meta:
            model = Movie
            fields = ['title']
    movie = MovieForReview(read_only=True)
    class Meta:
        model = Review
        fields = ['id',  'movie', 'title', 'content']
        read_only_fields = ['movie']
